requirements.txt
# Database libraries
pgcli
sqlalchemy
psycopg2

# Data processing libraries
pyarrow
pandas

# credentials management
python-dotenv